fix(java): consolidate canonicalization into provekit-ir (one BLAKE3 + JCS, not two); code-point key sort
Review fix for PR #593: the PR had added a second hand-rolled BLAKE3 + JCS encoder
to provekit-ir/Jcs.java when the Java tree already had a vetted one in
provekit-claim-envelope (BouncyCastle Blake3Digest, 'mirrors rust/csharp 1:1'),
and the new one sorted object keys by UTF-16 code unit (String.compareTo) rather
than Unicode code point (spec §7.3 / RFC 8785). Consolidated: moved the
BouncyCastle-backed BLAKE3 (Blake3.java) and the code-point-sorted JCS (Jcs.java)
into provekit-ir; deleted the duplicate copies from provekit-claim-envelope;
updated the Java consumers (claim-envelope, java-self-contracts, realize-java-core,
provekit-lift-java-source) to import the IR canonicalization classes; added IR
tests pinned to the Rust/b3sum canonical bytes + BLAKE3 vectors, including non-BMP
key ordering. mvn -DskipTests=false test passes across the affected + downstream
modules. (No menagerie/manifest.yaml change — that file is a destinations: list of
runnable exhibits, no *-language-signature catalog belongs there; #593 correctly
never added one.)
