+++ implementations/java/provekit-claim-envelope/src/main/java/com/provekit/claimenvelope/ProofEnvelope.java
@@ -30,6 +30,8 @@ import java.util.Map;
 import java.util.Objects;
 import java.util.TreeMap;
 
+import com.provekit.ir.Blake3;
+
 public final class ProofEnvelope {
 
     private ProofEnvelope() {}
