+++ implementations/java/provekit-claim-envelope/src/test/java/com/provekit/claimenvelope/Blake3Test.java
@@ -7,6 +7,8 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
 
+import com.provekit.ir.Blake3;
+
 class Blake3Test {
 
     @Test
