+++ implementations/java/provekit-claim-envelope/src/test/java/com/provekit/claimenvelope/BridgeV14RoundtripTest.java
@@ -40,6 +40,7 @@ import org.junit.jupiter.api.Test;
 import com.provekit.claimenvelope.ClaimEnvelope.BridgeTargetV14;
 import com.provekit.claimenvelope.ClaimEnvelope.MintBridgeV14Args;
 import com.provekit.claimenvelope.ClaimEnvelope.MintedEnvelope;
+import com.provekit.ir.Blake3;
 
 class BridgeV14RoundtripTest {
