+++ implementations/java/provekit-realize-java-core/pom.xml
@@ -14,6 +14,11 @@
     <dependencies>
         <dependency>
             <groupId>com.provekit</groupId>
+            <artifactId>provekit-ir</artifactId>
+            <version>0.1.0</version>
+        </dependency>
+        <dependency>
+            <groupId>com.provekit</groupId>
             <artifactId>provekit-claim-envelope</artifactId>
             <version>0.1.0</version>
         </dependency>
